chore(waiter-web): print bons on confirm, before payment (#131)
Reorder the waiter flow so the kitchen bon is printed as soon as the
waiter confirms the item selection, instead of at the payment step —
letting food prep and payment happen in parallel.

- MenuPage: the bottom CTA now submits + prints the order, shows the
  print result, then routes to the cart screen. The cart is no longer
  cleared on submit so its items carry over to be settled.
- OrderPage: becomes payment-only (Teilrechnung / split-bill + finish);
  the order-create/print logic is removed.
- Extract shared submit helpers to lib/order-submit.ts and the print
  result modal to components/PrintResultModal.tsx (used by MenuPage).
- Modal now renders the "queued" printer status (offline → queued bon).
